Moved queries into seperate files and update react router in App
