+++ src/components/GqlQueries/verbQueries.js
@@ -249,57 +249,6 @@ const LEVEL_SEVEN = gql`
   }
 `;
 
-const CREATE_USER = gql`
-  mutation($name: String!, $email: String!, $password: String!) {
-    createUser(data: { name: $name, email: $email, password: $password }) {
-      user {
-        name
-        email
-      }
-    }
-  }
-`;
-
-const LOGIN = gql`
-  mutation($email: String!, $password: String!) {
-    login(data: { email: $email, password: $password }) {
-      user {
-        name
-        email
-      }
-      token
-    }
-  }
-`;
-
-const LOGS = gql`
-  query {
-    logs {
-      id
-      verbInfinitive
-      tense
-      answer
-      correct
-      createdAt
-      user {
-        id
-        name
-        email
-      }
-    }
-  }
-`;
-
-const GET_USERS = gql`
-  query {
-    users {
-      id
-      name
-      email
-    }
-  }
-`;
-
 const verbQueries = [
   LEVEL_ONE,
   LEVEL_TWO,
@@ -310,4 +259,4 @@ const verbQueries = [
   LEVEL_SEVEN
 ];
 
-export { verbQueries, CREATE_USER, LOGIN, LOGS, GET_USERS };
+export { verbQueries };
